[MCJ-1367] refactor: auth/user 계층 패키지 구조 이동 및 import 경로 정리
