+++ src/main/kotlin/com/moongchijang/application/auth/dto/AccessTokenReissueResult.kt
@@ -1,0 +1,9 @@
+package com.moongchijang.application.auth.dto
+
+import com.moongchijang.application.auth.dto.response.AccessTokenResponse
+
+data class AccessTokenReissueResult(
+    val response: AccessTokenResponse,
+    val refreshToken: String,
+)
+
