+++ src/main/kotlin/com/moongchijang/application/auth/dto/AuthLoginResult.kt
@@ -1,0 +1,8 @@
+package com.moongchijang.application.auth.dto
+
+import com.moongchijang.application.auth.dto.response.AuthLoginResponse
+
+data class AuthLoginResult(
+    val response: AuthLoginResponse,
+    val refreshToken: String,
+)
