+++ src/main/kotlin/com/moongchijang/application/auth/dto/KakaoAuthUser.kt
@@ -1,0 +1,7 @@
+package com.moongchijang.application.auth.dto
+
+data class KakaoAuthUser(
+    val providerId: String,
+    val email: String,
+    val nickname: String,
+)
