+++ src/main/kotlin/com/moongchijang/application/auth/dto/request/KakaoLoginRequest.kt
@@ -1,0 +1,9 @@
+package com.moongchijang.application.auth.dto.request
+
+import jakarta.validation.constraints.NotBlank
+
+data class KakaoLoginRequest(
+    @field:NotBlank
+    var authorizationCode: String,
+    var redirectUri: String? = null,
+)
