+++ src/main/kotlin/com/moongchijang/application/auth/dto/response/AccessTokenResponse.kt
@@ -1,0 +1,7 @@
+package com.moongchijang.application.auth.dto.response
+
+data class AccessTokenResponse(
+    val accessToken: String,
+    val tokenType: String = "Bearer",
+    val expiresIn: Long,
+)
