+++ src/main/kotlin/com/moongchijang/application/user/dto/response/AdditionalInfoUpdatedResponse.kt
@@ -1,0 +1,26 @@
+package com.moongchijang.application.user.dto.response
+
+import com.moongchijang.domain.user.entity.User
+
+data class AdditionalInfoUpdatedResponse(
+    val id: Long,
+    val nickname: String,
+    val phoneNumber: String,
+    val signupCompleted: Boolean,
+) {
+    companion object {
+        fun from(user: User): AdditionalInfoUpdatedResponse {
+            val userId = user.id ?: error("user id must not be null")
+            val nickname = user.nickname ?: error("nickname must not be null")
+            val phoneNumber = user.phoneNumber ?: error("phoneNumber must not be null")
+
+            return AdditionalInfoUpdatedResponse(
+                id = userId,
+                nickname = nickname,
+                phoneNumber = phoneNumber,
+                signupCompleted = user.signupCompleted,
+            )
+        }
+    }
+}
+
