+++ src/main/kotlin/com/moongchijang/application/user/dto/response/NicknameAvailabilityResponse.kt
@@ -1,0 +1,7 @@
+package com.moongchijang.application.user.dto.response
+
+data class NicknameAvailabilityResponse(
+    val nickname: String,
+    val available: Boolean,
+)
+
