--- src/main/kotlin/com/moongchijang/domain/auth/application/dto/AccessTokenReissueResult.kt
@@ -1,9 +1,0 @@
-package com.moongchijang.domain.auth.application.dto
-
-import com.moongchijang.domain.auth.application.dto.response.AccessTokenResponse
-
-data class AccessTokenReissueResult(
-    val response: AccessTokenResponse,
-    val refreshToken: String,
-)
-
