--- src/main/kotlin/com/moongchijang/domain/auth/application/dto/AuthLoginResult.kt
@@ -1,8 +1,0 @@
-package com.moongchijang.domain.auth.application.dto
-
-import com.moongchijang.domain.auth.application.dto.response.AuthLoginResponse
-
-data class AuthLoginResult(
-    val response: AuthLoginResponse,
-    val refreshToken: String,
-)
