--- src/main/kotlin/com/moongchijang/domain/auth/application/dto/KakaoAuthUser.kt
@@ -1,7 +1,0 @@
-package com.moongchijang.domain.auth.application.dto
-
-data class KakaoAuthUser(
-    val providerId: String,
-    val email: String,
-    val nickname: String,
-)
