--- src/main/kotlin/com/moongchijang/domain/auth/application/dto/request/KakaoLoginRequest.kt
@@ -1,9 +1,0 @@
-package com.moongchijang.domain.auth.application.dto.request
-
-import jakarta.validation.constraints.NotBlank
-
-data class KakaoLoginRequest(
-    @field:NotBlank
-    var authorizationCode: String,
-    var redirectUri: String? = null,
-)
