--- src/main/kotlin/com/moongchijang/domain/auth/application/dto/response/AccessTokenResponse.kt
@@ -1,7 +1,0 @@
-package com.moongchijang.domain.auth.application.dto.response
-
-data class AccessTokenResponse(
-    val accessToken: String,
-    val tokenType: String = "Bearer",
-    val expiresIn: Long,
-)
