--- src/main/kotlin/com/moongchijang/domain/auth/application/dto/response/AuthLoginResponse.kt
@@ -1,9 +1,0 @@
-package com.moongchijang.domain.auth.application.dto.response
-
-data class AuthLoginResponse(
-    val accessToken: String,
-    val tokenType: String = "Bearer",
-    val expiresIn: Long,
-    val isNewUser: Boolean,
-    val user: AuthUserResponse,
-)
