--- src/main/kotlin/com/moongchijang/domain/auth/infrastructure/oauth/kakao/config/KakaoPropertiesConfig.kt
@@ -1,9 +1,0 @@
-package com.moongchijang.domain.auth.infrastructure.oauth.kakao.config
-
-import org.springframework.boot.context.properties.EnableConfigurationProperties
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-@EnableConfigurationProperties(KakaoProperties::class)
-class KakaoPropertiesConfig {
-}
