--- src/main/kotlin/com/moongchijang/domain/user/application/dto/request/AdditionalInfoUpsertRequest.kt
@@ -1,11 +1,0 @@
-package com.moongchijang.domain.user.application.dto.request
-
-import jakarta.validation.constraints.NotBlank
-
-data class AdditionalInfoUpsertRequest(
-    @field:NotBlank
-    val nickname: String,
-    @field:NotBlank
-    val phoneNumber: String,
-)
-
