--- src/main/kotlin/com/moongchijang/domain/user/application/dto/response/NicknameAvailabilityResponse.kt
@@ -1,7 +1,0 @@
-package com.moongchijang.domain.user.application.dto.response
-
-data class NicknameAvailabilityResponse(
-    val nickname: String,
-    val available: Boolean,
-)
-
