--- src/main/kotlin/com/moongchijang/domain/user/domain/entity/AuthProvider.kt
@@ -1,6 +1,0 @@
-package com.moongchijang.domain.user.domain.entity
-
-enum class AuthProvider {
-    KAKAO,
-    EMAIL
-}
