--- src/main/kotlin/com/moongchijang/domain/user/domain/entity/UserRole.kt
@@ -1,7 +1,0 @@
-package com.moongchijang.domain.user.domain.entity
-
-enum class UserRole {
-    BUYER,
-    SELLER,
-    ADMIN
-}
