+++ src/main/kotlin/com/moongchijang/domain/user/entity/AuthProvider.kt
@@ -1,0 +1,6 @@
+package com.moongchijang.domain.user.entity
+
+enum class AuthProvider {
+    KAKAO,
+    EMAIL
+}
