+++ src/main/kotlin/com/moongchijang/domain/user/entity/UserRole.kt
@@ -1,0 +1,7 @@
+package com.moongchijang.domain.user.entity
+
+enum class UserRole {
+    BUYER,
+    SELLER,
+    ADMIN
+}
