+++ src/main/kotlin/com/moongchijang/domain/user/repository/UserRepository.kt
@@ -1,0 +1,26 @@
+package com.moongchijang.domain.user.repository
+
+import com.moongchijang.domain.user.entity.AuthProvider
+import com.moongchijang.domain.user.entity.User
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface UserRepository : JpaRepository<User, Long> {
+
+    // 활성(미탈퇴) 유저 조회
+    fun findByProviderAndProviderIdAndDeletedAtIsNull(
+        provider: AuthProvider,
+        providerId: String,
+    ): User?
+
+    // 탈퇴 유저 조회
+    fun findByProviderAndProviderIdAndDeletedAtIsNotNull(
+        provider: AuthProvider,
+        providerId: String,
+    ): User?
+
+    fun existsByNicknameAndDeletedAtIsNull(nickname: String): Boolean
+
+    fun existsByEmailAndDeletedAtIsNull(email: String): Boolean
+
+    fun findByIdAndDeletedAtIsNull(id: Long): User?
+}
