+++ src/main/kotlin/com/moongchijang/infrastructure/oauth/kakao/config/KakaoPropertiesConfig.kt
@@ -1,0 +1,9 @@
+package com.moongchijang.infrastructure.oauth.kakao.config
+
+import org.springframework.boot.context.properties.EnableConfigurationProperties
+import org.springframework.context.annotation.Configuration
+
+@Configuration
+@EnableConfigurationProperties(KakaoProperties::class)
+class KakaoPropertiesConfig {
+}
