--- src/test/kotlin/com/moongchijang/domain/user/application/UserServiceTest.kt
@@ -1,123 +1,0 @@
-package com.moongchijang.domain.user.application
-
-import com.moongchijang.domain.user.domain.entity.AuthProvider
-import com.moongchijang.domain.user.domain.entity.User
-import com.moongchijang.domain.user.domain.repository.UserRepository
-import com.moongchijang.global.exception.CustomException
-import com.moongchijang.global.exception.ErrorCode
-import com.moongchijang.support.UserFixture.createKakaoUser
-import org.junit.jupiter.api.Assertions.assertEquals
-import org.junit.jupiter.api.Assertions.assertFalse
-import org.junit.jupiter.api.Assertions.assertTrue
-import org.junit.jupiter.api.Test
-import org.mockito.Mockito
-import org.mockito.Mockito.never
-import org.mockito.Mockito.verify
-import java.time.LocalDateTime
-
-class UserServiceTest {
-
-    private val userRepository: UserRepository = Mockito.mock(UserRepository::class.java)
-    private val userService = UserService(userRepository)
-
-    @Test
-    fun `활성 카카오 사용자가 있으면 기존 사용자 반환`() {
-        val existingUser = createKakaoUser(id = 1L, providerId = "kakao-1", nickname = "기존닉네임")
-
-        Mockito.`when`(
-            userRepository.findByProviderAndProviderIdAndDeletedAtIsNull(AuthProvider.KAKAO, "kakao-1"),
-        ).thenReturn(existingUser)
-
-        val (user, isNewUser) = userService.findOrCreateKakaoUser(
-            providerId = "kakao-1",
-            email = "test@example.com",
-            nickname = "신규닉네임",
-        )
-
-        assertEquals(1L, user.id)
-        assertFalse(isNewUser)
-        verify(userRepository, never()).save(Mockito.any(User::class.java))
-    }
-
-    @Test
-    fun `활성 사용자와 탈퇴 사용자가 없으면 신규 사용자 생성`() {
-        val savedUser = createKakaoUser(id = 10L, providerId = "kakao-new", email = "new@example.com", nickname = "신규유저")
-
-        Mockito.`when`(
-            userRepository.findByProviderAndProviderIdAndDeletedAtIsNull(AuthProvider.KAKAO, "kakao-new"),
-        ).thenReturn(null)
-        Mockito.`when`(
-            userRepository.findByProviderAndProviderIdAndDeletedAtIsNotNull(AuthProvider.KAKAO, "kakao-new"),
-        ).thenReturn(null)
-        Mockito.`when`(userRepository.save(Mockito.any(User::class.java))).thenReturn(savedUser)
-
-        val (user, isNewUser) = userService.findOrCreateKakaoUser(
-            providerId = "kakao-new",
-            email = "new@example.com",
-            nickname = "신규유저",
-        )
-
-        assertEquals(10L, user.id)
-        assertTrue(isNewUser)
-        verify(userRepository).save(Mockito.any(User::class.java))
-    }
-
-    @Test
-    fun `추가정보 입력 시 닉네임 형식이 잘못되면 예외`() {
-        val exception = org.junit.jupiter.api.assertThrows<CustomException> {
-            userService.updateAdditionalInfo(
-                userId = 1L,
-                nickname = "닉 네 임",
-                phoneNumber = "010-1234-5678",
-            )
-        }
-
-        assertEquals(ErrorCode.INVALID_NICKNAME_FORMAT, exception.errorCode)
-    }
-
-    @Test
-    fun `추가정보 입력 시 닉네임이 중복되면 예외`() {
-        val user = createKakaoUser(id = 1L, providerId = "kakao-dup", email = "dup@example.com", nickname = "기존닉네임")
-
-        Mockito.`when`(userRepository.findByIdAndDeletedAtIsNull(1L)).thenReturn(user)
-        Mockito.`when`(userRepository.existsByNicknameAndDeletedAtIsNull("중복닉네임")).thenReturn(true)
-
-        val exception = org.junit.jupiter.api.assertThrows<CustomException> {
-            userService.updateAdditionalInfo(
-                userId = 1L,
-                nickname = "중복닉네임",
-                phoneNumber = "010-1234-5678",
-            )
-        }
-
-        assertEquals(ErrorCode.DUPLICATE_NICKNAME, exception.errorCode)
-    }
-
-    @Test
-    fun `탈퇴일로부터 30일 미만이면 재가입 불가`() {
-        val deletedUser = createKakaoUser(
-            id = 20L,
-            providerId = "kakao-deleted",
-            email = "deleted@example.com",
-            nickname = "탈퇴유저",
-            deletedAt = LocalDateTime.now().minusDays(5),
-        )
-
-        Mockito.`when`(
-            userRepository.findByProviderAndProviderIdAndDeletedAtIsNull(AuthProvider.KAKAO, "kakao-deleted"),
-        ).thenReturn(null)
-        Mockito.`when`(
-            userRepository.findByProviderAndProviderIdAndDeletedAtIsNotNull(AuthProvider.KAKAO, "kakao-deleted"),
-        ).thenReturn(deletedUser)
-
-        val exception = org.junit.jupiter.api.assertThrows<CustomException> {
-            userService.findOrCreateKakaoUser(
-                providerId = "kakao-deleted",
-                email = "deleted@example.com",
-                nickname = "새닉네임",
-            )
-        }
-
-        assertEquals(ErrorCode.REJOIN_NOT_AVAILABLE_YET, exception.errorCode)
-    }
-}
